No longer try to create a thumbnail for our mp4 video.
We used to tell zencoder to create a thumbnail to go along with the
converted video, the same thumbnail youtube uses.  But youtube no
longer exposes the necessary information -- how many seconds into the
video is the frame it took as the thumbnail -- via its API.  (youtube
APIv2 is no more, sadness.)  So we can no longer create the thumbnail.

We *could* do something else like just pick a random time in, which
might be better than nothing, but I think we're planning to use custom
thumbnails one day soon-ish anyway, so we'll hold off on that.

Auditors: desmond

Test Plan:
Fingers crossed