Fix up logging for print_missing_converted_formats the *right* way.
I thought that the old code just wasn't calling addHandler(), so no
logging was happening at all, but it turns out it was just calling it
in an unexpected place.  I moved the addHandler() routine to util.py,
next to where the logger is defined, and now call it from
print_missing_converted_formats.py.

While in the area, I changed the logger to always log to stderr for
ERROR log messages.  I now use that so that errors are sent to
cron-mail (since cron-mail sends everything it sees on stdout or
stderr), rather than using 'print' directly.

Auditors: kamens

Test Plan:
Fingers crossed!  Will run it manually on youtube-export.