Improve print_missing_converted_formats to ignore new videos.
Summary:
If a video was *just* creatd, the converted formats often don't exist
yet, since it takes time to do the conversion.  So now we only
complain if a video is missing a conversion two days in a row.
Hopefully it takes less than 24 hours to do a conversion!

Test Plan:
Fingers crossed -- we'll see what output this run has tomorrow.
I will also check to make sure that the 'Saving converted-formats
output' message is run every day.

Reviewers: andy, chris

Reviewed By: chris

Subscribers: Jake

Differential Revision: https://phabricator.khanacademy.org/D26687