Change youtube-export to use the new webapp API.
Summary:
Since youtube-export is the only client of the old API, I'm replacing
it with a newer API that is more tailored to the needs of
youtube-export.  This will make it easier to maintain going forward.

I'll land this after landing the parallel webapp change that gets rid
of the old API and creates this new one.

Test Plan:
Mostly fingers crossed; we'll see how the next youtube-export run
goes.

Reviewers: tom, michaelpolyak

Reviewed By: michaelpolyak

Differential Revision: https://phabricator.khanacademy.org/D54634