Added logging for succesful video and script completion.
Summary: Added logging for successful video and script completion. These additional logs will be used to create charts and alerts around this script on StackDriver, so that we can catch issues with it early.

Test Plan: Run script locally and watch for correct logs. Update: I was unable to run the script locally due to issues setting up my google cloud credentials, but I was able to test by declaring the necessary variables and running the new lines of code in the python shell. Since we aren't changing any functionality, it should be sufficient to just make sure only these new lines work as expected.

Reviewers: neha

Reviewed By: neha

Subscribers: tom

Differential Revision: https://phabricator.khanacademy.org/D57263